requirements.txt
# Core ML and Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Framework and API
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# File Upload and Processing
werkzeug>=2.3.0

# Monitoring and Logging
psutil>=5.9.0

# Load Testing
locust>=2.15.0

# Development and Utilities
jupyter>=1.0.0
ipykernel>=6.25.0
python-dotenv>=1.0.0

# Image Processing
scikit-image>=0.21.0 